+++ js/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "shiny-dataframe-binding",
+  "name": "py-shiny-js",
   "private": true,
   "license": "MIT",
   "scripts": {
