+++ shiny/render/__init__.py
@@ -5,7 +5,7 @@ Tools for reactively rendering output for the user interface.
 from . import (  # noqa: F401
     transformer,  # pyright: ignore[reportUnusedImport]
 )
-from ._dataframe import (  # noqa: F401
+from ._data_frame import (  # noqa: F401
     CellPatch,  # pyright: ignore[reportUnusedImport]
     DataGrid,
     DataTable,
