+++ shiny/ui/dataframe/__init__.py
@@ -1,3 +1,3 @@
-from ._dataframe import output_data_frame
+from ._data_frame import output_data_frame
 
 __all__ = ("output_data_frame",)
