+++ tests/playwright/shiny/components/data_frame/edit/app.py
@@ -6,7 +6,7 @@
 from palmerpenguins import load_penguins_raw
 
 from shiny import App, Inputs, Outputs, Session, module, render, ui
-from shiny.render._dataframe import CellPatch
+from shiny.render._data_frame import CellPatch
 
 # TODO-barret-render.data_frame; Make an example that uses a dataframe that then updates a higher level reactive, that causes the df to update... which causes the table to render completely
 # TODO-barret-render.data_frame; When "updating" data, try to maintain the scroll, filter info when a new `df` is supplied;
