+++ tests/playwright/shiny/components/data_frame/validate_data_edit_mode/app.py
@@ -1,7 +1,7 @@
 from palmerpenguins import load_penguins_raw  # pyright: ignore[reportMissingTypeStubs]
 
 from shiny import App, Inputs, Outputs, Session, render, ui
-from shiny.render._dataframe import CellPatch
+from shiny.render._data_frame import CellPatch
 from shiny.types import SafeException
 
 app_ui = ui.page_fluid(
